Reference

MPSTimeEvolution.ExpValueCallback — Type

ExpValueCallback(
    operators, sites::Vector{<:Index}, measure_timestep::Float64
)

Construct an ExpValueCallback, providing some operators and a list of ITensor sites. The operator variable can be either a vector of LocalOperator objects, or a string (see parseoperators for instructions on the allowed syntax). Each operator will be measured on the associated sites during every step of the time evolution, and the results recorded inside the ExpValueCallback object as an ExpValueSeries for later analysis. The norm of the state, or the an equivalent quantity (trace, overlap of two sites...) where applicable, will also be computed after each step. The array sites is the same list of sites indices used to define MPSs and MPOs for the calculations.

MPSTimeEvolution.LocalMeasurementCallback — Method

LocalMeasurementCallback(ops::Vector{String},
                         sites::Vector{<:Index},
                         dt_measure::Float64)

Construct a LocalMeasurementCallback, providing an array ops of operator names which are strings recognized by the op function. Each of these operators will be measured on every site during every step of the time evolution, and the results recorded inside the object as a Measurement for later analysis. The array sites is the basis of sites used to define the MPS and MPO for the calculations.

MPSTimeEvolution.LocalOperator — Type

LocalOperator(terms::OrderedDict{Int,AbstractString})

A LocalOperator represents a product of local operators, whose names (strings, as recognized by ITensors) are specified by factors and acting on sites which are not necessarily consecutive. For example, the operator $1 ⊗ A ⊗ 1 ⊗ C$ would be represented as {2 => "A", 4 => "C"}.

MPSTimeEvolution.LocalPosMeasurementCallback — Method

LocalPosMeasurementCallback(ops::Vector{opPos},
                            sites::Vector{<:Index},
                            dt_measure::Float64)

Construct a LocalPosMeasurementCallback, providing an array ops of opPos objects which represent operators associated to specific sites. Each of these operators will be measured on the given site during every step of the time evolution, and the results recorded inside the LocalPosMeasurementCallback object as a Measurement for later analysis. The array sites is the basis of sites used to define the MPS and MPO for the calculations.

MPSTimeEvolution.LocalPosVecMeasurementCallback — Method

LocalPosVecMeasurementCallback(
    ops::Vector{opPos}, sites::Vector{<:Index}, dt_measure::Float64
)

Like LocalPosMeasurementCallback, but with vectorized operators.

MPSTimeEvolution.Measurement — Type

A Measurement object is an alias for `Vector{Float64}`, in other words an
array of real numbers.

Given a Measurement `M`, the result for the measurement at step `n` is `M[n]`.

MPSTimeEvolution.NoTEvoCallback — Type

NoTEvoCallback is a trivial implementation of an evolution callback (<:TEvoCallback)
object.

MPSTimeEvolution.SuperfermionCallback — Type

SuperfermionCallback(
    operators, sites::Vector{<:Index}, measure_timestep::Float64
)

Construct a SuperfermionCallback providing some operators and a list of ITensor sites. The operator variable can be either a vector of LocalOperator objects, or a string (see parseoperators for instructions on the allowed syntax). Each operator will be measured on the given site during every step of the time evolution, and the results recorded inside the SuperfermionCallback object as an ExpValueSeries for later analysis. The array sites is the same basis of sites used to define the MPS and MPO for the calculations.

This struct is defined the same say as ExpValueCallback, but it allows using Julia's multiple dispatch features to choose the correct method to measure expectation values.

MPSTimeEvolution.TEvoCallback — Type

A TEvoCallback can implement the following methods:

apply!(cb::TEvoCallback, psi ; t, kwargs...): apply the callback with the

current state psi (e.g. perform some measurement)

checkdone!(cb::TEvoCallback, psi; t, kwargs...): check whether some criterion

to stop the time evolution is satisfied (e.g. convergence of cbervable, error too large) and return true if so.

callback_dt(cb::TEvoCallback): time-steps at which the callback needs access

for the wave-function (e.g. for measurements). This is used for TEBD evolution where several time-steps are bunched together to reduce the cost.

MPSTimeEvolution.opPos — Type

opPos(op::String, pos::Integer)

An opPos object is an operator op (a string recognised by ITensors' op function) attached to a specific site pos.

MPSTimeEvolution.adaptbonddimensions! — Method

function adaptbonddimensions!(
    v::MPS, PH::AbstractProjMPO, max_bond::Int, convergence_factor_bonddims::Real
)

If necessary, enlarge the bond dimensions of the MPS v, so that it can dynamically and inhomogeneously restructure itself as the complexity of the correlations grow during the time evolution generated by H [1].

References:

[1] Dunnett, Angus J. and Chin, Alex W. (2020). “Dynamically Evolving Bond-Dimensions within the one-site Time-Dependent-Variational-Principle method for Matrix Product States: Towards efficient simulation of non-equilibrium open quantum dynamics” https://doi.org/10.48550/arXiv.2007.13528

MPSTimeEvolution.adaptiveadjtdvp1vec! — Function

adaptiveadjtdvp1vec!(
    [solver,] operator, initialstate::MPS, L, dt, tmax, meas_stride; kwargs...
)
adaptiveadjtdvp1vec!(
    [solver,] operator, initialstates::Vector{MPS}, L, dt, tmax, meas_stride; kwargs...
)

Like adjtdvp1vec!, but grows the bond dimensions of the MPS of the operator along the time evolution until a certain convergence criterium is met. The keyword argument convergence_factor_bonddims, which defaults to 1e-4, controls the convergence of the adaptation algorithm.

For an explanation of the other arguments, see adjtdvp1vec!.

MPSTimeEvolution.adaptivetdvp1! — Method

adaptivetdvp1!([solver,] state::MPS, H::MPO, dt, tmax; kwargs...)
adaptivetdvp1!([solver,] state::MPS, H::Vector{MPO}, dt, tmax; kwargs...)

Like tdvp1!, but grows the bond dimensions of the MPS along the time evolution until a certain convergence criterium is met. The keyword argument convergence_factor_bonddims, which defaults to 1e-4, controls the convergence of the adaptation algorithm.

For an explanation of the other arguments, see tdvp1!.

MPSTimeEvolution.adaptivetdvp1vec! — Method

adaptivetdvp1vec!(solver, state, L, dt, tmax; kwargs...)

Like tdvp1vec!, but grows the bond dimensions of the MPS along the time evolution until a certain convergence criterium is met. The keyword argument convergence_factor_bonddims, which defaults to 1e-4, controls the convergence of the adaptation algorithm.

For an explanation of the other arguments, see tdvp1vec!.

MPSTimeEvolution.adj — Method

adj(x)

Returns the adjoint (conjugate transpose) of x. It can be an ITensor, an MPS or an MPO. Note that it is not the same as ITensors.adjoint.

MPSTimeEvolution.adjtdvp1vec! — Function

adjtdvp1vec!(
    [solver,] operator, initialstate::MPS, L, dt, tmax, meas_stride; kwargs...
)
adjtdvp1vec!(
    [solver,] operator, initialstates::Vector{MPS}, L, dt, tmax, meas_stride; kwargs...
)

Integrate the equation of motion $d/dt Xₜ = φ(Xₜ)$ using the one-site TDVP algorithm, from 0 to tmax in time steps of dt. The MPS operator represents $X₀$ in a vectorised form, while L is an MPO representing the operator $φ$ that evolves $X₀$ in time. The expectation value of the operator is periodically computed (each meas_stride) on the initial state initialstate or on a list initialstates of initial states.

See tdvp1vec! for a list of possible keyword arguments.

Other keyword arguments (specific to this function)

initialstatelabels: labels for the initial states (to mark the expectation values in the output files). Defaults to ["1", "2", …, "N"] where N is the amount of initial states provided to the function.

MPSTimeEvolution.apply! — Method

apply!(cb::LocalMeasurementCallback, psi; t, sweepend, sweepdir, bond, alg, kwargs...)

Calculates the expectation values of the operators stored in cb on state, if the conditions appropriate to the evolution algorithm are met.

MPSTimeEvolution.apply! — Method

apply!(
    cb::LocalPosMeasurementCallback, state; t, sweepend, sweepdir, bond, alg, kwargs...
)

Calculates the expectation values of the operators stored in cb on state, if the conditions appropriate to the evolution algorithm are met.

MPSTimeEvolution.apply! — Method

apply!(
    cb::LocalPosVecMeasurementCallback, state;
    t, sweepend, sweepdir, bond, alg, kwargs...
)

Calculates the expectation values of the operators stored in cb on state, if the conditions appropriate to the evolution algorithm are met.

MPSTimeEvolution.bondconvergencemeasure — Method

bondconvergencemeasure(PH::AbstractProjMPO, state::MPS, bond::Integer)

Return a measure of the convergence for the bond dimension on the bond (bond, bond+1) of the MPS state for a time-evolution determined by PH. See [1] for details.

References:

[1] Dunnett, Angus J. and Chin, Alex W. (2020). “Dynamically Evolving Bond-Dimensions within the one-site Time-Dependent-Variational-Principle method for Matrix Product States: Towards efficient simulation of non-equilibrium open quantum dynamics” https://doi.org/10.48550/arXiv.2007.13528

MPSTimeEvolution.enlargelinks — Method

enlargelinks(v, dims; ref_state)

Increase the bond dimensions of the MPS v to dims by adding an all-zero MPS to it, with a direct sum. The result is an MPS orthogonalized on the first site whose maximum bond dimensions are dims, except possibly at its edges, where the bond dimension is naturally lower.

In order to obtain the all-zero MPS, a random MPS with the given bond dimensions is computed, so that the correct link structure is maintained; it is then multiplied by zero and added to v.

If the MPS has some conserved quantum numbers, then a product state must also be supplied as the keyword argument ref_state, because the random MPS is obtained by randomising this initial state (which also determines the total QN of the resulting random MPS). This argument may take the form of one of the many available ways to build an MPS from the MPS(sites, ...) function, such as a string, an array of strings, or a function.

Examples

Without quantum numbers:

julia> N = 9; s = siteinds("Fermion", N);

julia> v = MPS(s, "Emp");

julia> v_ext = enlargelinks(v, 10; ref_state="Emp");

julia> print(linkdims(v_ext))
[2, 4, 8, 10, 10, 8, 4, 2]

With quantum numbers:

julia> N = 9; s = siteinds("Fermion", N; conserve_nfparity=true);

julia> v = MPS(s, n -> isodd(n) ? "Occ" : "Emp");

julia> v_ext = enlargelinks(v, 10; ref_state=n -> isodd(n) ? "Occ" : "Emp");

julia> print(linkdims(v_ext))
[1, 2, 4, 8, 10, 8, 4, 2]

MPSTimeEvolution.expvalues — Function

expvalues(cb::ExpValueCallback)
expvalues(cb::ExpValueCallback, lop::LocalOperator)
expvalues(cb::ExpValueCallback, name::AbstractString)

expvalues(cb::SuperfermionCallback)
expvalues(cb::SuperfermionCallback, lop::LocalOperator)
expvalues(cb::SuperfermionCallback, name::AbstractString)

Retrieve the expectation values of the operators stored in the callback cb. The time series of an individual operator lop within cb can be directly accessed by expvalues(cb, lop), where lop is either a LocalOperator or a string that defines a single LocalOperator.

MPSTimeEvolution.growMPS! — Method

growMPS!(v::MPS, d::Integer)

Grow the dimension of v's bond indices (in place) so that the result has a bond dimension of d on each of its bonds. Return the overlap of the new MPS with the original one.

MPSTimeEvolution.growMPS! — Method

growMPS!(v::MPS, dims::Vector{<:Integer})

Grow the dimension of v's bond indices (in place) so that the result has a bond dimension of dims[n] on bond (n, n+1). Return the overlap of the new MPS with the original one.

MPSTimeEvolution.growMPS — Method

growMPS(v::MPS, d::Integer)

Grow the dimension of v's bond indices so that the result has a bond dimension of d on each of its bonds. Return the new MPS and its overlap with the original one.

MPSTimeEvolution.growMPS — Method

growMPS(v::MPS, dims::Vector{<:Integer})

Grow the dimension of v's bond indices so that the result has a bond dimension of dims[n] on bond (n, n+1). Return the new MPS and its overlap with the original one.

MPSTimeEvolution.growbond! — Method

growbond!(v::MPS, bond::Integer; increment::Integer=1)

Grow the bond dimension of the MPS v between sites bond and bond+1 by increment. Return the new bond dimension.

MPSTimeEvolution.isoncurrentbond — Method

isoncurrentbond(op::opPos, i::Int, alg)

Return whether the local operator op is associated to bond i (or bonds i and i+1, if the algorithm alg is 2-site TDVP).

MPSTimeEvolution.jointtdvp1! — Function

jointtdvp1!([solver,] states::Tuple{MPS, MPS}, H::MPO, dt, tmax; kwargs...)
jointtdvp1!([solver,] states::Tuple{MPS, MPS}, H::Vector{MPO}, dt, tmax; kwargs...)

Integrate the Schrödinger equation $d/dt ψⁱₜ = -i H ψⁱₜ$ using the one-site TDVP algorithm, for both states $ψ¹$ and $ψ²$ in states in parallel, from 0 to tmax in time steps of dt. The evolution operator can be given either as a single MPO or as a vector of MPOs, in the latter case the evolution operator is taken to be the sum of the elements in the vector.

Instead of returning the step-by-step expectation value of operators on a single state, it returns quantities of the form $⟨ψ¹ₜ,Aψ²ₜ⟩$ for each operator $A$ specified in the given callback object, together with the overlap $⟨ψ¹ₜ,ψ²ₜ⟩$.

Other arguments

solver: a function which takes three arguments A, t, B (and possibly other keyword arguments) where t is a time step, B an ITensor and A a linear operator on B, returning the time-evolved B. It defaults to KrylovKit.exponentiate.

dt: time step of the evolution.

tmax: end time of the evolution.

Optional keyword arguments

cb: a callback object describing the observables.

hermitian (default: true): whether H is an Hermitian operator.

exp_tol (default: 1e-14): accuracy per unit time for KrylovKit.exponentiate.

krylovdim (default: 30): maximum dimension of the Krylov subspace that will be constructed.

maxiter (default: 100): number of times the Krylov subspace can be rebuilt.

normalize (default: true): whether the states are renormalised after each step.

io_file (default: nothing): output file for step-by-step measurements.

io_ranks (default: nothing): output file for step-by-step bond dimensions.

io_times (default: nothing): output file for simulation wall-clock times.

store_psi0 (default: false): whether to keep information about the initial state.

which_decomp (default: "qr"): name of the decomposition method for the sweeps.

progress (default: true): whether to display a progress bar during the evolution.

MPSTimeEvolution.maxlinkdims — Method

maxlinkdims(st::AbstractString, N[, maxbonddim])

Return a vector containing the maximum bond dimensions that an MPS of site type st and length N can have, taking into account both the physical dimension of its sites and (if provided) a manually set upper bound maxbonddim.

MPSTimeEvolution.maxlinkdims — Method

maxlinkdims(v::MPS[, maxbonddim])

Return a vector containing the maximum bond dimensions that the MPS v can have, taking into account both the physical dimension of its sites and (if provided) a manually set upper bound maxbonddim. The function assumes that all sites share the same physical dimension.

MPSTimeEvolution.meanordefault — Function

meanordefault(v, default=nothing)

Compute the mean of v, unless default is given, in which case return default.

MPSTimeEvolution.measure_localops! — Method

measure_localops!(cb::ExpValueCallback, state::MPS, site::Int, alg::TDVP1)

Measure on the MPS state each operator whose support starts on site defined inside the callback object cb.

MPSTimeEvolution.measure_localops! — Method

measure_localops!(cb::ExpValueCallback, ψ₁::MPS, ψ₂::MPS, alg::TDVP1)

Compute the inner product $⟨ψ₁, Aψ₂⟩$ for each operator $A$, defined in the callback object $cb$.

MPSTimeEvolution.measure_localops! — Method

measure_localops!(cb::ExpValueCallback, ψ::MPS, alg::TDVP1vec)

Measure each operator defined inside the callback object cb on the state ψ.

MPSTimeEvolution.measure_localops! — Method

function measure_localops!(cb::LocalMeasurementCallback, wf::ITensor, i::Int)

Measure each operator defined inside the callback object cb on the tensor wf, which represents the sit i of a bigger MPS.

This requires that the MPS where wf comes from has been properly orthogonalized such that its orthogonality center is right on site i. This is understood when calculating expectation values using this function; otherwise, it gives unexpected results.

MPSTimeEvolution.measure_localops! — Method

measure_localops!(
    cb::LocalPosMeasurementCallback,
    wf::ITensor,
    i::Int,
    alg
)

Measure each operator defined inside the callback object cb on the given tensor wf at the selected bond or site i (depending on the algorithm alg).

MPSTimeEvolution.measure_localops! — Method

measure_localops!(cb::LocalPosMeasurementCallback, psi::MPS, i::Int, alg)

Measure each operator defined inside the callback object cb on the given MPS psi at the selected site i (depending on the algorithm alg).

The function is implemented so as to contract the minimum number of tensor sites as needed to provide a correct result, given i and the orthocentre of the MPS.

MPSTimeEvolution.measure_localops! — Method

measure_localops!(
    cb::LocalPosVecMeasurementCallback,
    ψ::MPS,
    i::Int,
    alg
)

Measure each operator defined inside the callback object cb on the MPS ψ at the selected bond or site i (depending on the algorithm alg).

Use this version with vectorized operators.

MPSTimeEvolution.measure_localops! — Method

measure_localops!(cb::SuperfermionCallback, ψ::MPS, alg::TDVP1vec)

Measure each operator defined inside the callback object cb on the state ψ.

MPSTimeEvolution.mpo — Method

mpo(sites::Vector{<:Index}, l::LocalOperator)

Return an MPO that represents the operator described by l.

MPSTimeEvolution.mps — Method

mps(sites::Vector{<:Index}, l::LocalOperator)

Return an MPS that represents the operator described by l in a vectorised form. The MPS will have the factors in l on the domain of the operator, or Id outside of the domain.

MPSTimeEvolution.parseoperators — Method

parseoperators(s::AbstractString)

Parse the string s as a list of LocalOperator objects, with the following rules:

operators are written as products of local operators of the form name(site) where name is a string and site is an integer;

operators on different sites can be multiplied together by writing them one after the other, i.e. a(1)b(2);

for convenience, writing a comma-separated list of numbers in the parentheses expands to a list of operators with the same name on each of the sites in the list, i.e. a(1,2,3) is interpreted as a(1),a(2),a(3).

Example

julia> parseoperators("x(1)y(3),y(4),z(1,2,3)")
5-element Vector{LocalOperator}:
 x(1)y(3)
 y(4)
 z(1)
 z(2)
 z(3)

MPSTimeEvolution.parseoperators — Method

parseoperators(d::Dict)

Translate a dictionary into a list of LocalOperator objects, where each (key, value) pair is interpreted as the operator key (a string) on each of the sites contained in value (a list of integers), separately. This type of input does not support products of operators on different sites.

Example

julia> ops = Dict("x" => [1, 2, 3], "y" => [2]);

julia> parseoperators(ops)
4-element Vector{LocalOperator}:
 x{1}
 x{2}
 x{3}
 y{2}

MPSTimeEvolution.recompute! — Method

recompute!(P::AbstractProjMPO, psi::MPS, n::Int)

Recompute P's projection operators assuming that psi has changed on sites (n, n+1). The position of the projection is not changed.

MPSTimeEvolution.recompute! — Method

recompute!(P::ProjMPOSum, psi::MPS, n::Int)

Recompute P's projection operators assuming that psi has changed on sites (n, n+1). The position of the projection is not changed.

MPSTimeEvolution.smart_contract — Method

smart_contract(A::opPos, ψ::MPS, site_range)

Return the expectation value $⟨ψ|A|ψ⟩$ contracting only sites in the given range. It is assumed that ψ's orthocentre lies within site_range.

MPSTimeEvolution.tdvp1! — Function

tdvp1!([solver,] state::MPS, H::Vector{MPO}, dt, tmax; kwargs...)
tdvp1!([solver,] state::MPS, H::MPO, dt, tmax; kwargs...)

Integrate the Schrödinger equation $d/dt ψₜ = -i H ψₜ$ using the one-site TDVP algorithm, where state is an MPS representing the state of the system. The Hamiltonian H can be given either as a single MPO or as a vector of MPOs, in the latter case the total Hamiltonian is taken to be the sum of the elements in the vector.

Other arguments

solver: a function which takes three arguments A, t, B (and possibly other keyword arguments) where t is a time step, B an ITensor and A a linear operator on B, returning the time-evolved B. It defaults to KrylovKit.exponentiate.

dt: time step of the evolution.

tmax: end time of the evolution.

Optional keyword arguments

callback: a callback object describing the observables.

hermitian (default: true): whether H is an Hermitian operator.

exp_tol (default: 1e-14): accuracy per unit time for KrylovKit.exponentiate.

krylovdim (default: 30): maximum dimension of the Krylov subspace that will be constructed.

maxiter (default: 100): number of times the Krylov subspace can be rebuilt.

normalize (default: true): whether state is renormalised after each step.

io_file (default: nothing): output file for step-by-step measurements.

io_ranks (default: nothing): output file for step-by-step bond dimensions.

io_times (default: nothing): output file for simulation wall-clock times.

store_psi0 (default: false): whether to keep information about the initial state.

which_decomp (default: "qr"): name of the decomposition method for the sweeps.

progress (default: true): whether to display a progress bar during the evolution.

MPSTimeEvolution.tdvp1vec! — Function

tdvp1vec!([solver,] state::MPS, L::MPO, dt, tmax; kwargs...)
tdvp1vec!([solver,] state::MPS, L::Vector{MPO}, dt, tmax; kwargs...)

Integrate the equation of motion $d/dt ρₜ = L(ρₜ)$ using the one-site TDVP algorithm, from 0 to tmax in time steps of dt. The MPS state represents $ρ$ in a vectorised form. The evolution operator can be given either as a single MPO or as a vector of MPOs, in the latter case the evolution operator is taken to be the sum of the elements in the vector.

Other arguments

solver: a function which takes three arguments A, t, B (and possibly other keyword arguments) where t is a time step, B an ITensor and A a linear operator on B, returning the time-evolved B. It defaults to KrylovKit.exponentiate.

dt: time step of the evolution.

tmax: end time of the evolution.

Optional keyword arguments

callback: a callback object describing the observables.

hermitian (default: false): whether L is an Hermitian operator.

exp_tol (default: 1e-14): accuracy per unit time for KrylovKit.exponentiate.

krylovdim (default: 30): maximum dimension of the Krylov subspace that will be constructed.

maxiter (default: 100): number of times the Krylov subspace can be rebuilt.

normalize (default: false): whether state is renormalised after each step.

io_file (default: nothing): output file for step-by-step measurements.

io_ranks (default: nothing): output file for step-by-step bond dimensions.

io_times (default: nothing): output file for simulation wall-clock times.

store_psi0 (default: false): whether to keep information about the initial state.

which_decomp (default: "qr"): name of the decomposition method for the sweeps.

progress (default: true): whether to display a progress bar during the evolution.

MPSTimeEvolution.tdvp2! — Method

tdvp2!(ψ, H::MPO, timestep, endtime; kwargs...)

Evolve the MPS ψ up to time endtime using the two-site time-dependent variational principle as described in [1].

Keyword arguments:

All keyword arguments controlling truncation which are accepted by ITensors.replacebond!, namely:

maxdim::Int: If specified, keep only maxdim largest singular values after

applying the gate.

mindim::Int: Minimal number of singular values to keep if truncation is performed

according to value specified by cutoff.

cutoff::Float: If specified, keep the minimal number of singular values such that the

discarded weight is smaller than cutoff (but the bond dimension will be kept smaller than maxdim).

absoluteCutoff::Bool: If true truncate all singular-values whose square is smaller

than cutoff.

In addition the following keyword arguments are supported:

hermitian::Bool (true) : whether the MPO H represents an Hermitian operator.

This will be passed to the Krylov exponentiation routine (KrylovKit.exponentiate) which will in turn use a Lancosz algorithm in the case of an hermitian operator.

exp_tol::Float (1e-14) : The error tolerance for KrylovKit.exponentiate (note that

default value was not optimized yet, so you might want to play around with it).

progress::Bool (true) : If true a progress bar will be displayed

References:

[1] Haegeman, J., Lubich, C., Oseledets, I., Vandereycken, B., & Verstraete, F. (2016). “Unifying time evolution and optimization with matrix product states” Physical Review B, 94(16). https://doi.org/10.1103/PhysRevB.94.165116

MPSTimeEvolution.tdvp2vec! — Method

tdvp2vec!(state, L::MPO, dt, tmax; kwargs...)

Evolve the MPS state using the MPO H from 0 to tmax using an integration step dt.

MPSTimeEvolution.tdvpMC! — Method

tdvpMC!(state, H::MPO, dt, tmax; kwargs...)

Evolve the MPS state using the MPO H from 0 to tmax using an integration step dt.

MPSTimeEvolution.tdvp_site_update! — Method

function tdvp_site_update!(
    solver, PH, state::MPS, i::Int, dt;
    sweepdir, current_time, which_decomp="qr", hermitian, exp_tol, krylovdim, maxiter
)

Update site i of the MPS state using the 1-site TDVP algorithm with time step dt. The keyword argument sweepdir indicates the direction of the current sweep.

MPSTimeEvolution.to_vec — Method

to_vec(x)

Transform x into a Vector, and return the vector, and a closure which inverts the transformation.

Modeled after FiniteDifferences.to_vec:

https://github.com/JuliaDiff/FiniteDifferences.jl/blob/main/src/to_vec.jl

MPSTimeEvolution.writeheaders_data — Method

writeheaders_data(io_file, cb; kwargs...)

Prepare the output file io_file, writing the column headers for storing the data of the observables defined in cb, the time steps, and other basic quantities.

MPSTimeEvolution.writeheaders_ranks — Method

writeheaders_ranks(ranks_file, N)

Prepare the output file ranks_file, writing the column headers for storing the data relative to the ranks of a MPS of the given length N.

MPSTimeEvolution.writeheaders_stime — Method

writeheaders_stime(times_file)

Prepare the output file times_file, writing the column headers for the simulation time data.